docs: Add application structure design diagram
Add an application structure design diagram depicting the relationship between classes, functions, and page events in the application. Also included are major class attributes.

Note that the DesignFile class is not completely designed at this time since it is an optional feature. Its design will be completed when this feature becomes required.